+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.1.2
+
+* Added `.iter()`, `.iter_mut()`, `.iter_top()`, `.keys()`, and `.keys_top()` methods to `ScopeMap`
+* Added `.iter()` and `.iter_top()` methods to `ScopeSet`
+
 ### 0.1.1
 
 * Added `.define_parent()` method to `ScopeMap` and `ScopeSet`